requirements.txt
https://github.com/Rapptz/discord.py/tarball/async#egg=package-0.10.0a0
requests
PyYaml
mcstatus
